Use dynamic import to load menu
