+++ src/js/burokku.js
@@ -11,8 +11,8 @@ doc.classList.replace('no-js', 'js')
 
 initEvents()
 initBlocks()
-initWallet()
-initMenu()
+const wallet = initWallet()
+initMenu({ show: () => wallet.coins })
 initGame()
 initColorSchemes()
 initServiceWorker()
